requirements.txt
streamlit==1.39.0
flask==3.0.3
requests==2.32.3
googletrans==4.0.0-rc1
python-dotenv==1.0.1
google-generativeai==0.8.3
pycountry==24.6.1
requests>=2.25.1